configs: Add config file for 'test' option impact evaluation
This commit adds a config file, namely 'hint_test'.  It has two 100 MiB
memory regions and two phases.  In each phase, it access almost only one
region with 99 % probability.  However, the frequently accessed region
is changed by phase.  In other words, it is designed to have big benefit
from the hint.  Evaluation result on a test machine is as below:
```
 # echo $((110*1024*1024)) > ./memory.limit_in_bytes
 # /masim/masim /masim/configs/hint_test
 example phase 1:                      25,179 accesses/msec, 5013 msecs
 example phase 2:                      22,911 accesses/msec, 5000 msecs
 # /home/sjpark/masim/masim -t madvise /masim/configs/hint_test
 example phase 1:                      24,138 accesses/msec, 5001 msecs
 example phase 2:                      24,579 accesses/msec, 5018 msecs
 # /masim/masim -t mlock /masim/configs/hint_test
 example phase 1:                      58,877 accesses/msec, 5000 msecs
 example phase 2:                      59,454 accesses/msec, 5000 msecs
```

Signed-off-by: SeongJae Park <sj38.park@gmail.com>